% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tldr-package.R
\name{tldrPages}
\alias{tldrPages}
\alias{tldrPages-package}
\title{tldrPages: Short-form documentation in the console}
\description{
A package that implements quick "tldr"-style documentation in the R console,
providing the \code{tldr()} function for users and the roxygen2 extension
\code{tldr-roclet} for package developers.
}
